+++ src/ewokssphinx/pydantic_utils.py
@@ -1,7 +1,6 @@
-from __future__ import annotations
-
 import importlib
-from typing import Any, Callable
+from collections.abc import Callable
+from typing import Any
 
 from docutils import nodes
 from docutils.nodes import Node