+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.0.0
 
-- **⚠️ Breaking change:** Python 3.8 was dropped
+- **⚠️ Breaking change:** No longer supports Python 3.8 and Python 3.9
 - ✨ Add rich display of inputs `ewokstasks` directive when specified with a pydantic model: type annotation, description and examples
 - ✨ `ewokstasks` directive now displays inputs and outputs as lists rather than on a single line